mage: Organize package into separate files per concern
